+++ webview-ui/src/components/welcome/WelcomeView.tsx
@@ -6,7 +6,7 @@ import { vscode } from "../../utils/vscode"
 import ApiOptions from "../settings/ApiOptions"
 
 const WelcomeView = () => {
-	const { apiConfiguration } = useExtensionState()
+	const { apiConfiguration, currentApiConfigName } = useExtensionState()
 
 	const [errorMessage, setErrorMessage] = useState<string | undefined>(undefined)
 
@@ -17,11 +17,15 @@ const WelcomeView = () => {
 			return
 		}
 		setErrorMessage(undefined)
-		vscode.postMessage({ type: "apiConfiguration", apiConfiguration })
+		vscode.postMessage({
+			type: "upsertApiConfiguration",
+			text: currentApiConfigName,
+			apiConfiguration,
+		})
 	}
 
 	return (
-		<div style={{ position: "fixed", top: 0, left: 0, right: 0, bottom: 0, padding: "0 20px" }}>
+		<div className="flex flex-col min-h-screen px-0 pb-5">
 			<h2>Hi, I'm Roo!</h2>
 			<p>
 				I can do all kinds of tasks thanks to the latest breakthroughs in agentic coding capabilities and access
@@ -32,12 +36,13 @@ const WelcomeView = () => {
 
 			<b>To get started, this extension needs an API provider.</b>
 
-			<div style={{ marginTop: "10px" }}>
+			<div className="mt-3">
 				<ApiOptions fromWelcomeView />
-				<div style={{ display: "flex", flexDirection: "column", gap: "5px" }}>
-					<VSCodeButton onClick={handleSubmit} style={{ marginTop: "3px" }}>
-						Let's go!
-					</VSCodeButton>
+			</div>
+
+			<div className="sticky bottom-0 bg-[var(--vscode-editor-background)] py-3">
+				<div className="flex flex-col gap-1.5">
+					<VSCodeButton onClick={handleSubmit}>Let's go!</VSCodeButton>
 					{errorMessage && <span className="text-destructive">{errorMessage}</span>}
 				</div>
 			</div>